also move ServerConnectionManager to connection module
